fix ValueError: high is out of bounds for int32

+++ stable_audio_tools/inference/generation.py
@@ -129,7 +129,7 @@ def generate_diffusion_cond(
         
     # Seed
     # The user can explicitly set the seed to deterministically generate the same output. Otherwise, use a random seed.
-    seed = seed if seed != -1 else np.random.randint(0, 2**32 - 1)
+    seed = seed if seed != -1 else np.random.randint(0, 2**32 - 1, dtype=np.uint32)
     print(seed)
     torch.manual_seed(seed)
     # Define the initial noise immediately after setting the seed
